fixed product crud and complaint crud issues

+++ GroceryStore/Controllers/Customer/Customer_ComplaintsController.cs
@@ -8,6 +8,7 @@ using System.Web;
 using System.Web.Mvc;
 using GroceryStore.Controllers.Dto;
 using GroceryStore.Models;
+using GroceryStore.Utilities;
 
 namespace GroceryStore.Controllers
 {
@@ -18,13 +19,16 @@ namespace GroceryStore.Controllers
         // GET: Customer_Complaints
         public ActionResult Index()
         {
-            if (Session["RoleName"] == null)
+            if (Session["RoleName"] == null || Session["RoleName"].ToString() != AppRoles.Customer)
             {
                 return RedirectToAction("Auth", "Auth");
             }
-            int id = Int32.Parse(Session["userID"].ToString());
-            var user = db.Users.Where(x => x.UserID == id).FirstOrDefault();
-            var customer_Complaints = db.Customer_Complaints.Include(c => c.Complaint_Type).Include(c => c.User).Where(m => m.Customer_Id == id) ;
+
+            int u_id = Int32.Parse(Session["userID"].ToString());
+            var user = db.Users.Where(x => x.UserID == u_id).FirstOrDefault();
+            var customer = db.Customers.Where(x => x.UserID == user.UserID).FirstOrDefault();
+
+            var customer_Complaints = db.Customer_Complaints.Include(c => c.Complaint_Type).Where(m => m.Customer_Id == customer.Customer_id) ;
 
             return View(new UserDto {  Customer_Complaints= customer_Complaints.ToList(), User = user });
         }
@@ -47,16 +51,14 @@ namespace GroceryStore.Controllers
         // GET: Customer_Complaints/Create
         public ActionResult Create()
         {
-            ViewBag.Complaint_type_Id = new SelectList(db.Complaint_Type, "Complaint_Type_Id", "Complaint_Tpe");
-            ViewBag.Customer_Id = new SelectList(db.Customers, "Customer_id", "UserID");
-
             if (Session["RoleName"] == null)
             {
                 return RedirectToAction("Auth", "Auth");
             }
+
+            ViewBag.Complaint_type_Id = new SelectList(db.Complaint_Type, "Complaint_Type_Id", "Complaint_Tpe");
             int id = Int32.Parse(Session["userID"].ToString());
             var user = db.Users.Where(x => x.UserID == id).FirstOrDefault();
-
             return View(new UserDto {User = user }); 
         }
 
@@ -65,23 +67,41 @@ namespace GroceryStore.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Create([Bind(Include = "Complaint_Id,Customer_Id,Complaint_type_Id,Complaint,Complaint_Status,date")] Customer_Complaints customer_Complaints)
+        public ActionResult Create([Bind(Include = "Complaint_type_Id, Complaint")] Customer_Complaints customer_Complaints)
         {
-            if (ModelState.IsValid)
+            if (Session["RoleName"] == null || Session["RoleName"].ToString() != AppRoles.Customer)
             {
-                db.Customer_Complaints.Add(customer_Complaints);
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                return RedirectToAction("Auth", "Auth");
             }
 
-            ViewBag.Complaint_type_Id = new SelectList(db.Complaint_Type, "Complaint_Type_Id", "Complaint_Tpe", customer_Complaints.Complaint_type_Id);
-            ViewBag.Customer_Id = new SelectList(db.Customers, "Customer_id", "Customer_id", customer_Complaints.Customer_Id);
-            return View(customer_Complaints);
+            int u_id = Int32.Parse(Session["userID"].ToString());
+            var user = db.Users.Where(x => x.UserID == u_id).FirstOrDefault();
+            var customer = db.Customers.Where(x => x.UserID == user.UserID).FirstOrDefault();
+
+            customer_Complaints.Customer_Id = customer.Customer_id;
+            customer_Complaints.date = DateTime.Now;
+            customer_Complaints.Complaint_Status = "PENDING";
+            customer_Complaints.Complaint = Request.Form["customer_Complaint.Complaint"].ToString();
+            db.Customer_Complaints.Add(customer_Complaints);
+            db.SaveChanges();
+            
+            
+            return RedirectToAction("Index");
+      
         }
 
         // GET: Customer_Complaints/Edit/5
         public ActionResult Edit(int? id)
         {
+            if (Session["RoleName"] == null || Session["RoleName"].ToString() != AppRoles.Customer)
+            {
+                return RedirectToAction("Auth", "Auth");
+            }
+
+            int u_id = Int32.Parse(Session["userID"].ToString());
+            var user = db.Users.Where(x => x.UserID == u_id).FirstOrDefault();
+            var customer = db.Customers.Where(x => x.UserID == user.UserID).FirstOrDefault();
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -91,9 +111,10 @@ namespace GroceryStore.Controllers
             {
                 return HttpNotFound();
             }
+
             ViewBag.Complaint_type_Id = new SelectList(db.Complaint_Type, "Complaint_Type_Id", "Complaint_Tpe", customer_Complaints.Complaint_type_Id);
-            ViewBag.Customer_Id = new SelectList(db.Customers, "Customer_id", "Customer_id", customer_Complaints.Customer_Id);
-            return View(customer_Complaints);
+            
+            return View(new UserDto { User = user , customer_Complaint = customer_Complaints});
         }
 
         // POST: Customer_Complaints/Edit/5
@@ -101,17 +122,31 @@ namespace GroceryStore.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "Complaint_Id,Customer_Id,Complaint_type_Id,Complaint,Complaint_Status,date")] Customer_Complaints customer_Complaints)
+        public ActionResult Edit()
         {
-            if (ModelState.IsValid)
+            if (Session["RoleName"] == null || Session["RoleName"].ToString() != AppRoles.Customer)
             {
-                db.Entry(customer_Complaints).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                return RedirectToAction("Auth", "Auth");
             }
-            ViewBag.Complaint_type_Id = new SelectList(db.Complaint_Type, "Complaint_Type_Id", "Complaint_Tpe", customer_Complaints.Complaint_type_Id);
-            ViewBag.Customer_Id = new SelectList(db.Customers, "Customer_id", "Customer_id", customer_Complaints.Customer_Id);
-            return View(customer_Complaints);
+
+            int u_id = Int32.Parse(Session["userID"].ToString());
+            var user = db.Users.Where(x => x.UserID == u_id).FirstOrDefault();
+            var customer = db.Customers.Where(x => x.UserID == user.UserID).FirstOrDefault();
+
+            int complaint_id = Int32.Parse(Request.Form["customer_Complaint.Complaint_Id"]);
+
+            var cust_instance = db.Customer_Complaints.Where(
+                x => x.Customer_Id == customer.Customer_id && x.Complaint_Id == complaint_id
+                ).FirstOrDefault();
+
+
+            if (cust_instance.Complaint_Status.Contains("PENDING")) {
+                cust_instance.Complaint = Request.Form["customer_Complaint.Complaint"].ToString();
+                db.Entry(cust_instance).State = EntityState.Modified;
+                db.SaveChanges();
+            }
+           
+            return RedirectToAction("Index");
         }
 
         // GET: Customer_Complaints/Delete/5
@@ -126,7 +161,11 @@ namespace GroceryStore.Controllers
             {
                 return HttpNotFound();
             }
-            return View(customer_Complaints);
+            int u_id = Int32.Parse(Session["userID"].ToString());
+            var user = db.Users.Where(x => x.UserID == u_id).FirstOrDefault();
+            var customer = db.Customers.Where(x => x.UserID == user.UserID).FirstOrDefault();
+
+            return View(new UserDto { User = user, customer_Complaint = customer_Complaints });
         }
 
         // POST: Customer_Complaints/Delete/5
@@ -134,9 +173,18 @@ namespace GroceryStore.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+
+            if (Session["RoleName"] == null || Session["RoleName"].ToString() != AppRoles.Customer)
+            {
+                return RedirectToAction("Auth", "Auth");
+            }
+
             Customer_Complaints customer_Complaints = db.Customer_Complaints.Find(id);
-            db.Customer_Complaints.Remove(customer_Complaints);
-            db.SaveChanges();
+
+            if (customer_Complaints.Complaint_Status.Contains("PENDING")) { 
+                db.Customer_Complaints.Remove(customer_Complaints);
+                db.SaveChanges();
+            }
             return RedirectToAction("Index");
         }
